Store Tests: fixed circular import and a bug
